Drop the Sainsmart TFT
Its precompiled disk image was broken in a few ways that killed this project.  
- Wheezy was older than I wanted, 
- apt-get update broke the tft driver and 
- downgrading to a 5 year old version of Python (3.2.3) started requiring a whole other set of workarounds.  
None of these were unsurmountable, but it was a lot of uphill effort just to get to a neutral point.  There are better RPi monitors out there, but cell phone pairing is probably even cleaner.  And it might give an opportunity to write an iphone app.